GP-10789 Fix campaign not being propagated from activity
This fixes an issue where campaign_id was not being propagated
from the contract activity to the recurring contribution during
Contract.process_scheduled_modifications API calls. This was due
to the campaign_id parameter not being requested when fetching
scheduled contract activities.

Tests were extended to cover the expected behaviour of campaign
propagation.