GP-10618 Add ability to synchronize with ContributionRecur
This adds a new Contract.synchronize_contribution_recur API which allows
the synchronization of data from recurring contributions to the contract
payment information fields. This allows changes made to recurring
contributions outside of CE to be fed back into CE data structures.

The implementation was done in a pseudo change class which, unlike
all other change types, is not actually backed by an activity, but
rather just used ad-hoc via the API.